% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers2words.R
\name{numbers2words}
\alias{numbers2words}
\title{Converts a numeric value to letters based number}
\usage{
numbers2words(x)
}
\arguments{
\item{input}{An arbitrary number}
}
\description{
Converts a numeric value to letters based number
}
